refactor: 임시로 만든 EditCard2를 EditCard 파일로 변경, ScheduleCard에서 EditCard 생성하는 부분 삭제
