+++ js/components/schedule/scheduleEditCard.js
@@ -1,148 +1,162 @@
 import { TEXT_LENGTH_LIMIT } from "../../utils.js";
+import {
+    getScheduleCardDataById,
+    updateScheduleCard,
+} from "../model/scheduleModel.js";
 
-export class ScheduleEditCard {
-    constructor({ original, passedEventHandler }) {
-        this.$originalCard = original;
-        this.$originalCardTitle = this.$originalCard.querySelector(
-            ".schedule-card__title"
-        );
-        this.$originalCardBody = this.$originalCard.querySelector(
-            ".schedule-card__body"
-        );
-        this.passedEventHandler = passedEventHandler;
-        this.$editCard;
-        this.init();
+let $originalCard;
+let $originalCardTitle;
+let $originalCardBody;
+
+let cardId;
+let columnId;
+
+let $editCard;
+let $editCardTitle;
+let $editCardBody;
+let $editCardEditBtn;
+let $editCardCancleBtn;
+
+let isShow = false;
+
+export const scheduleEditCardInit = ({ target }) => {
+    if (isShow) {
+        return;
     }
 
-    init() {
-        this.createEditCard();
-        this.render();
-        this.setEvent();
+    $originalCard = target.closest(".schedule-card");
+    if (!$originalCard) {
+        return;
     }
 
-    setDOMElement() {
-        this.$editCardTitle = this.$editCard.querySelector(
-            ".schedule-edit-card__title"
-        );
-        this.$editCardBody = this.$editCard.querySelector(
-            ".schedule-edit-card__body"
-        );
-        this.$editCardEditBtn = this.$editCard.querySelector(
-            ".schedule-edit-card__edit-btn"
-        );
-        this.$editCardCancleBtn = this.$editCard.querySelector(
-            ".schedule-edit-card__cancel-btn"
-        );
+    setOrigianlCardValue();
+
+    $editCard = createEditCard();
+    setEditCardValue();
+
+    render();
+    setEvent();
+};
+
+const setOrigianlCardValue = () => {
+    $originalCardTitle = $originalCard.querySelector(".schedule-card__title");
+    $originalCardBody = $originalCard.querySelector(".schedule-card__body");
+
+    cardId = $originalCard.dataset.cardId;
+    columnId = $originalCard.closest(".schedule-column").dataset.id;
+};
+
+const setEditCardValue = () => {
+    $editCardTitle = $editCard.querySelector(".schedule-edit-card__title");
+    $editCardBody = $editCard.querySelector(".schedule-edit-card__body");
+    $editCardEditBtn = $editCard.querySelector(".schedule-edit-card__edit-btn");
+    $editCardCancleBtn = $editCard.querySelector(
+        ".schedule-edit-card__cancel-btn"
+    );
+};
+
+const render = () => {
+    const parentNode = $originalCard.parentNode;
+    parentNode.replaceChild($editCard, $originalCard);
+    isShow = true;
+};
+
+const setEvent = () => {
+    $editCard.addEventListener("click", clickEventHandler);
+    $editCard.addEventListener("input", inputEventHandler);
+};
+
+const toggleEditBtn = (booleanValue) => {
+    if (booleanValue) {
+        $editCardEditBtn.classList.replace("inactive", "active");
+    } else {
+        $editCardEditBtn.classList.replace("active", "inactive");
+    }
+};
+
+const adjustInputHeight = ($input) => {
+    $input.style.height = `auto`;
+    $input.style.height = `${$input.scrollHeight}px`;
+};
+
+const inputEventHandler = ({ target }) => {
+    if (target === $editCardTitle || target === $editCardBody) {
+        const cardTitle = $editCardTitle.value;
+
+        toggleEditBtn(cardTitle);
+        adjustInputHeight($editCardTitle);
+        adjustInputHeight($editCardBody);
+    }
+};
+
+const cancelEdit = () => {
+    const parentNode = $editCard.parentNode;
+    parentNode.replaceChild($originalCard, $editCard);
+    isShow = false;
+};
+
+const editBtnClickEventHandler = (target) => {
+    if (target.classList.contains("inactive")) {
+        return;
     }
 
-    render() {
-        const parentNode = this.$originalCard.parentNode;
-        parentNode.replaceChild(this.$editCard, this.$originalCard);
+    const cardData = {
+        title: $editCardTitle.value,
+        body: $editCardBody.value,
+        caption: "author by web",
+        id: cardId,
+    };
+    updateScheduleCard(columnId, cardData);
+
+    $originalCardTitle.innerText = cardData.title;
+    $originalCardBody.innerText = cardData.body;
+
+    const parentNode = $editCard.parentNode;
+    parentNode.replaceChild($originalCard, $editCard);
+    isShow = false;
+};
+
+const clickEventHandler = ({ target }) => {
+    if (target === $editCardCancleBtn) {
+        cancelEdit();
     }
-
-    setEvent() {
-        this.$editCard.addEventListener(
-            "click",
-            this.clickEventHandler.bind(this)
-        );
-        this.$editCard.addEventListener(
-            "input",
-            this.inputEventHandler.bind(this)
-        );
+    if (target === $editCardEditBtn) {
+        editBtnClickEventHandler(target);
     }
+};
 
-    inputEventHandler({ target }) {
-        if (target === this.$editCardTitle || target === this.$editCardBody) {
-            const cardTitle = this.$editCardTitle.value;
+const createEditCard = () => {
+    const cardData = getScheduleCardDataById(columnId, cardId);
 
-            this.toggleEditBtn(cardTitle, this.$editCardEditBtn);
-            this.adjustInputHeight(this.$editCardTitle);
-            this.adjustInputHeight(this.$editCardBody);
-        }
-    }
+    $editCard = document.createElement("div");
+    $editCard.classList.add("schedule-edit-card");
+    $editCard.dataset.cardId = cardId;
+    $editCard.innerHTML = template(cardData.title, cardData.body);
 
-    toggleEditBtn(booleanValue, $editBtn) {
-        if (booleanValue) {
-            $editBtn.classList.replace("inactive", "active");
-        } else {
-            $editBtn.classList.replace("active", "inactive");
-        }
-    }
+    return $editCard;
+};
 
-    adjustInputHeight($input) {
-        $input.style.height = `auto`;
-        $input.style.height = `${$input.scrollHeight}px`;
-    }
-
-    clickEventHandler({ target }) {
-        if (target === this.$editCardCancleBtn) {
-            this.cancelEdit();
-        }
-        if (target === this.$editCardEditBtn) {
-            this.editBtnClickEventHandler(target);
-        }
-    }
-
-    editBtnClickEventHandler(target) {
-        if (target.classList.contains("inactive")) {
-            return;
-        }
-
-        const cardId = this.$editCard.dataset.cardId;
-        const cardData = {
-            title: this.$editCardTitle.value,
-            body: this.$editCardBody.value,
-            caption: "author by web",
-            id: cardId,
-        };
-        this.passedEventHandler.updateCard(cardData);
-
-        this.$originalCardTitle.innerText = cardData.title;
-        this.$originalCardBody.innerText = cardData.body;
-
-        const parentNode = this.$editCard.parentNode;
-        parentNode.replaceChild(this.$originalCard, this.$editCard);
-    }
-
-    cancelEdit() {
-        const parentNode = this.$editCard.parentNode;
-        parentNode.replaceChild(this.$originalCard, this.$editCard);
-    }
-
-    createEditCard() {
-        const cardId = this.$originalCard.dataset.cardId;
-        const cardData = this.passedEventHandler.getCardData(cardId);
-
-        this.$editCard = document.createElement("div");
-        this.$editCard.classList.add("schedule-edit-card");
-        this.$editCard.dataset.cardId = cardId;
-        this.$editCard.innerHTML = this.template(cardData.title, cardData.body);
-
-        this.setDOMElement();
-    }
-
-    template(cardTitle, cardBody) {
-        return `<form class="schedule-edit-card__text-container">
-                    <textarea 
-                        class="schedule-edit-card__title"  
-                        placeholder="제목을 입력하세요"
-                        rows="1"
-                        maxLength="${TEXT_LENGTH_LIMIT}"
-                    >${cardTitle}</textarea>
-                    <textarea 
-                        class="schedule-edit-card__body" 
-                        placeholder="내용을 입력하세요"
-                        rows="1"
-                        maxLength="${TEXT_LENGTH_LIMIT}"
-                    >${cardBody}</textarea>
-                </form>
-                <div class="schedule-edit-card__btns-container">
-                    <button class="schedule-edit-card__cancel-btn">
-                        취소
-                    </button>
-                    <button class="schedule-edit-card__edit-btn active">
-                        수정
-                    </button>
-                </div>`;
-    }
-}
+const template = (cardTitle, cardBody) => {
+    return `<form class="schedule-edit-card__text-container">
+                <textarea 
+                    class="schedule-edit-card__title"  
+                    placeholder="제목을 입력하세요"
+                    rows="1"
+                    maxLength="${TEXT_LENGTH_LIMIT}"
+                >${cardTitle}</textarea>
+                <textarea 
+                    class="schedule-edit-card__body" 
+                    placeholder="내용을 입력하세요"
+                    rows="1"
+                    maxLength="${TEXT_LENGTH_LIMIT}"
+                >${cardBody}</textarea>
+            </form>
+            <div class="schedule-edit-card__btns-container">
+                <button class="schedule-edit-card__cancel-btn">
+                    취소
+                </button>
+                <button class="schedule-edit-card__edit-btn active">
+                    수정
+                </button>
+            </div>`;
+};
