+++ js/index.js
@@ -1,7 +1,7 @@
 import { ScheduleColumn } from "./components/schedule/scheduleColumn.js";
 import { mouseDownEventHandler } from "./components/schedule/scheduleCardDrag.js";
 import { scheduleModel } from "./components/model/scheduleModel.js";
-import { scheduleEditCardInit } from "./components/schedule/scheduleEditCard2.js";
+import { scheduleEditCardInit } from "./components/schedule/scheduleEditCard.js";
 
 const $main = document.querySelector("#main");
 const scheduleColumns = [];
